WIP - Service Insertion - changes required for VPNaaS
Implements the VPN specific changes required for the service base
and insertion spec @ I595d7fe224f49f8cbc4203569d7862deaa74e1e7
which has been approved for Juno.

Note that this is early WIP, in particular it needs to be rebased on
the definition of ServiceBase and the ServiceInterface db models
(the latter is carried here for now) when this becomes available.

TODO_LIST:
1. DB Migrations
2. Tests